requirements.txt
streamlit>=1.31.0
aiohttp>=3.9.3
python-dotenv>=1.0.0
nest-asyncio>=1.5.8
openai>=1.12.0
